+++ examples/src/lib.rs
@@ -1,43 +1,18 @@
-use chrono::prelude::*;
-use std::time::{Duration, Instant};
-use tm1637_gpio_driver::gpio_api::setup_wiringpi;
-use tm1637_gpio_driver::{TM1637Adapter, DisplayState, Brightness};
-use tm1637_gpio_driver::mappings::SpecialCharBits;
-use tm1637_gpio_driver::fourdigit7segdis::{display_current_time_in_loop, display_stopwatch, STOPWATCH_MAX, display_timer};
 use std::ops::Add;
+use std::time::{Duration, Instant};
+use chrono::Local;
+use tm1637_gpio_driver::{Brightness, DisplayState, TM1637Adapter};
+use tm1637_gpio_driver::fourdigit7segdis::{display_current_time_in_loop, display_stopwatch, display_timer, STOPWATCH_MAX};
+use tm1637_gpio_driver::mappings::SpecialCharBits;
 
 // We have 4 displays
-const DISPLAYS_COUNT: usize = 4;
+pub const DISPLAYS_COUNT: usize = 4;
 
 /// One second in Micro seconds.
-const SECOND: u64 = 1E6 as u64;
+pub const SECOND: u64 = 1E6 as u64;
 
-/// Simple example that shows you how you can use the driver along with crate "wiringpi" to display
-/// content on the 4-digit 7-segment display by AZDelivery.
-/// This is the display shown in the gif of the readme.
-///
-/// This demo shows multiple kinds of using the display to show data.
-fn main() {
-    // use any GPIO pin you want. This is the number of the pin on the board.
-    // The numbers in the example here are available on the Raspberry Pi for example.
-    let clk_pin = 18;
-    let dio_pin = 23;
+pub fn run_demo(mut tm1637display: TM1637Adapter) {
 
-    // setup
-    // 100µs should be totally save; less could work; depends on cable length and other factors
-    // high frequencies are tricky; it even worked with zero waiting for me.. but better be safe!
-    let bit_delay_fn = || sleep_busy_waiting(10);
-    let bit_delay_fn = Box::from(bit_delay_fn);
-    let mut tm1637display = setup_wiringpi(clk_pin, dio_pin, bit_delay_fn);
-
-    // display "1 2 3 4"
-    let data: [u8; DISPLAYS_COUNT] = [
-        TM1637Adapter::encode_digit(1),
-        TM1637Adapter::encode_digit(2),
-        TM1637Adapter::encode_digit(3),
-        TM1637Adapter::encode_digit(4),
-    ];
-    tm1637display.write_segments_raw(&data, 0);
     sleep_busy_waiting(SECOND);
 
     // ##############################################################################
@@ -120,7 +95,7 @@ fn main() {
 /// Sleeps/waits actively for x µs. Doesn't send the thread to sleep.
 /// This solution is way more accurate when it comes to times <= 100µs.
 /// Because context switches of the threads is too slow on Raspberry Pi (3).
-fn sleep_busy_waiting(us: u64) {
+pub fn sleep_busy_waiting(us: u64) {
     let target_time = Instant::now().add(Duration::from_micros(us));
     loop {
         if Instant::now() >= target_time {