+++ src/lib.rs
@@ -17,6 +17,7 @@
 
 // from rust core library; no "external crate" in the manner that this is no crates.io dependency;
 // needed because no_std
+#[macro_use] // wee need the !format macro
 extern crate alloc;
 
 // Import our enums/arrays for the symbol mappings to the 7 segment display
@@ -44,16 +45,6 @@ use crate::mappings::{NumCharBits, UpCharBits, SpecialCharBits, LoCharBits};
 /// 7-segment display from AzDelivery only uses 4.
 pub const DISPLAY_REGISTERS_COUNT: usize = 6;
 
-
-/// Mode of GPIO Pins.
-#[repr(u8)]
-pub enum GpioPinMode {
-    /// Input-Pin.
-    INPUT,
-    /// Output-Pin.
-    OUTPUT,
-}
-
 /// The value of a GPIO pin.
 #[repr(u8)]
 #[derive(Clone, Copy)]
@@ -80,8 +71,6 @@ impl From<u8> for GpioPinValue {
 pub struct TM1637Adapter {
     /// Function that writes the value on the GPIO pin that acts as the clock.
     pin_clock_write_fn: Box<dyn Fn(GpioPinValue)>,
-    /// Function that changes the mode of the data in and out pin.
-    pin_dio_mode_fn: Box<dyn Fn(GpioPinMode)>,
     /// Function that writes the value on the GPIO pin that acts as data in and out.
     pin_dio_write_fn: Box<dyn Fn(GpioPinValue)>,
     /// Function that reads from the data in and out pin.
@@ -171,22 +160,25 @@ pub enum ISA {
 }
 
 impl TM1637Adapter {
-    pub fn new(pin_clock_mode_fn: Box<dyn Fn(GpioPinMode)>,
-               pin_clock_write_fn: Box<dyn Fn(GpioPinValue)>,
-               pin_dio_mode_fn: Box<dyn Fn(GpioPinMode)>,
+    /// Creates a new object to interact via GPIO with a TM1637.
+    ///
+    /// * `pin_clock_write_fn` function to write bit to CLK pin
+    /// * `pin_dio_write_fn` function to write bit to DIO pin
+    /// * `pin_dio_read_fn` function to read value from DIO pin
+    /// * `bit_delay_fn` function that is invoked after a bit has been written to a pin.
+    ///                  Probably 1 or even 0 µs are fine. This is just to be sure. It depends
+    ///                  on your hardware and your GPIO driver.
+    pub fn new(pin_clock_write_fn: Box<dyn Fn(GpioPinValue)>,
                pin_dio_write_fn: Box<dyn Fn(GpioPinValue)>,
                pin_dio_read_fn: Box<dyn Fn() -> GpioPinValue>,
                bit_delay_fn: Box<dyn Fn() -> ()>) -> TM1637Adapter {
 
         // assume both are already output pins - this is the contract that needs to be fulfilled!
-        (pin_clock_mode_fn)(GpioPinMode::OUTPUT);
-        (pin_dio_mode_fn)(GpioPinMode::OUTPUT);
         (pin_clock_write_fn)(GpioPinValue::LOW);
         (pin_dio_write_fn)(GpioPinValue::LOW);
 
         TM1637Adapter {
             pin_clock_write_fn,
-            pin_dio_mode_fn,
             pin_dio_write_fn,
             pin_dio_read_fn,
             bit_delay_fn,
@@ -461,9 +453,6 @@ impl TM1637Adapter {
         self.bit_delay();
         (self.pin_clock_write_fn)(GpioPinValue::HIGH);
 
-        // prepare read
-        (self.pin_dio_mode_fn)(GpioPinMode::INPUT);
-        self.bit_delay();
         let ack: GpioPinValue = (self.pin_dio_read_fn)();
 
         // wait a few cycles for ACK to be more fail safe
@@ -477,7 +466,6 @@ impl TM1637Adapter {
             }
         }
 
-        (self.pin_dio_mode_fn)(GpioPinMode::OUTPUT);
         (self.pin_clock_write_fn)(GpioPinValue::LOW);
         (self.pin_dio_write_fn)(GpioPinValue::LOW);
         self.bit_delay();